Refactor and splitting classes
