+++ EnergyApiFunction/Classes/AccountDetails.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Energy_Consumption_Function.Classes
+{
+    public class AccountDetails
+    {
+        public string number { get; set; }
+        public Property[] properties { get; set; }
+    }
+}
