+++ EnergyApiFunction/OctopusConsumptionFunction.cs
@@ -18,25 +18,17 @@ namespace Energy_Consumption_Function
         private readonly ILogger<OctopusConsumptionFunction> _logger;
         private readonly HttpClient _client;
 
-
         public OctopusConsumptionFunction(ILogger<OctopusConsumptionFunction> logger, HttpClient client)
         {
             _logger = logger;
             _client = client;
-
         }
 
         [Function("ConsumptionFunction")]
         public async Task<HttpResponseData> Run([HttpTrigger(AuthorizationLevel.Function, "get")] HttpRequestData req)
-
-
         {
-            _logger.LogInformation("C# HTTP trigger function processed a request.");
-
-
-            var common = new Common();
-
-
+            _logger.LogInformation($"{nameof(OctopusConsumptionFunction)} processed a request.");
+            var common = new Common(_client, _logger);
 
             string dateFrom = req.Query["from"];
             string dateTo = req.Query["to"];
@@ -56,48 +48,46 @@ namespace Energy_Consumption_Function
 
             if (string.IsNullOrEmpty(dateFrom) || string.IsNullOrEmpty(dateTo) || string.IsNullOrEmpty(energyType))
             {
-                return FormatResponse(req, HttpStatusCode.BadRequest, "Please pass a date range and energy type on the query string or in the request body");
+                return HelperFunctions.FormatResponse(req, HttpStatusCode.BadRequest, "Please pass a date range and energy type on the query string or in the request body");
             }
 
             try
             {
-                var account = await common.GetAccountDetails(_logger);
+                var account = await common.GetAccountDetails();
 
                 var dateFromUtc = DateTime.Parse(dateFrom).Date;
                 var dateToUtc = DateTime.Parse(dateTo).Date;
 
-                var elecTariff = common.GetAgreementCost(account, dateFromUtc, dateToUtc, "electricity-tariffs");                  
+                var elecTariff = HelperFunctions.GetAgreementCost(account, dateFromUtc, dateToUtc, "electricity-tariffs");                  
 
-                var gasTariff =  common.GetAgreementCost(account, dateFromUtc, dateToUtc, "gas-tariffs");
+                var gasTariff =  HelperFunctions.GetAgreementCost(account, dateFromUtc, dateToUtc, "gas-tariffs");
 
-                var tariff = await common.GetTariff<ElecTariff>(dateFrom, dateTo, elecTariff.FirstOrDefault().tariff_code, "electricity-tariffs", _logger);
+                var tariff = await common.GetTariff<ElecTariff>(dateFrom, dateTo, elecTariff.FirstOrDefault().tariff_code, "electricity-tariffs");
 
-                var tariffGas = await common.GetTariff<GasTariff>(dateFrom, dateTo, gasTariff.FirstOrDefault().tariff_code,  "gas-tariffs", _logger);
+                var tariffGas = await common.GetTariff<GasTariff>(dateFrom, dateTo, gasTariff.FirstOrDefault().tariff_code,  "gas-tariffs");
 
                 var dd = tariff.results.Where(x => x.payment_method == "DIRECT_DEBIT").FirstOrDefault();    
 
                 var ddGas = tariffGas.results.Where(x => x.payment_method == "DIRECT_DEBIT").FirstOrDefault();
 
                 _logger.LogInformation("Calling GetEnergyConsumption: From: {0}, {1}", dateFrom, dateTo);
-                var response = await common.GetEnergyConsumption(dateFrom, dateTo, _logger);
+                var response = await common.GetEnergyConsumption(dateFrom, dateTo);
 
-                var join = response.results.Where(x => x.interval_start.Date >= dd.valid_from.Date)
-                    .Select(x => new
-                    {x.interval_start, x.consumption, x.interval_end, dd.value_exc_vat, dd.value_inc_vat
-                    }).ToList();
+                var join = HelperFunctions.JoinConsumptionTariffs(response, dd);
 
                 _logger.LogInformation("Calling GetGasConsumption: From: {0}, {1}", dateFrom, dateTo);
-                var gasResponse = await common.GetGasConsumption(dateFrom, dateTo, _logger);
+                var gasResponse = await common.GetGasConsumption(dateFrom, dateTo);
 
-                var gasJoin = gasResponse.results.Where(x => x.interval_start.Date >= ddGas.valid_from.Date)
-                    .Select(x => new
-                    {x.interval_start, x.consumption, x.interval_end, ddGas.value_exc_vat, ddGas.value_inc_vat
-                                       }).ToList();
+                var gasJoin = HelperFunctions.JoinConsumptionTariffs(gasResponse, ddGas);              
 
-                var mergedResponse = new { Electricity = join, Gas =gasJoin };
+                var mergedResponse = new RequestResponse
+                {
+                    Electricity = join,
+                    Gas = gasJoin
+                };  
                 _logger.LogInformation("Response: {0}", JsonConvert.SerializeObject(mergedResponse));
                 string responseMessage = JsonConvert.SerializeObject(mergedResponse);
-                return FormatResponse(req, HttpStatusCode.OK, responseMessage);
+                return HelperFunctions.FormatResponse(req, HttpStatusCode.OK, responseMessage);
 
 
             }
@@ -105,20 +95,11 @@ namespace Energy_Consumption_Function
             {
                 _logger.LogError(ex.Message);
 
-                return FormatResponse(req, HttpStatusCode.BadRequest, "An error occured");
+                return HelperFunctions.FormatResponse(req, HttpStatusCode.BadRequest, "An error occured");
             }
         }
 
-        public static HttpResponseData FormatResponse(HttpRequestData req, HttpStatusCode HttpStatusCode, string message)
-        {
-           var responseMessageData = req.CreateResponse(HttpStatusCode);
-           responseMessageData.Headers.Add("Content-Type", "application/json");
-           responseMessageData.WriteString(message);
-           return responseMessageData;
-        }
-
-
-
+       
     }
